tailored-offers-demo/requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0

# UI
streamlit>=1.40.0

# Data handling
pandas>=2.0.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
